Create a single downloadable sample.
