+++ samples/snippets/csharp/System.DateTime/Program.cs
@@ -1,0 +1,15 @@
+﻿namespace SystemDateTimeReference
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            Instantiation.Snippets();
+            StringFormat.Snippets();
+            Parsing.Snippets();
+            Resolution.Snippets();
+            CalendarSamples.Snippets();
+            Persistence.Snippets();
+        }
+    }
+}
