--- samples/snippets/csharp/VS_Snippets_CLR_System/system.datetime.parsing/cs/parsingex1.cs
@@ -1,125 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Globalization;
-using System.Threading;
-
-public class Example
-{
-   public static void Main()
-   {
-// <Snippet1>
-      Thread.CurrentThread.CurrentCulture = CultureInfo.CreateSpecificCulture("en-GB");
-
-      DateTime date1 = new DateTime(2013, 6, 1, 12, 32, 30);
-      List<string> badFormats = new List<String>();
-      
-      Console.WriteLine($"{"Date String",-37} {"Date",-19}\n");
-      foreach (var dateString in date1.GetDateTimeFormats()) {
-         DateTime parsedDate;
-         if (DateTime.TryParse(dateString, out parsedDate))
-            Console.WriteLine($"{dateString,-37} {DateTime.Parse(dateString),-19}");
-         else
-            badFormats.Add(dateString);
-      } 
-      
-      // Display strings that could not be parsed.
-      if (badFormats.Count > 0) {
-         Console.WriteLine("\nStrings that could not be parsed: ");
-         foreach (var badFormat in badFormats)
-            Console.WriteLine($"   {badFormat}");         
-// The example displays the following output:
-//       Date String                           Date               
-//       
-//       01/06/2013                            01/06/2013 00:00:00
-//       01/06/13                              01/06/2013 00:00:00
-//       1/6/13                                01/06/2013 00:00:00
-//       1.6.13                                01/06/2013 00:00:00
-//       2013-06-01                            01/06/2013 00:00:00
-//       01 June 2013                          01/06/2013 00:00:00
-//       1 June 2013                           01/06/2013 00:00:00
-//       01 June 2013 12:32                    01/06/2013 12:32:00
-//       01 June 2013 12:32                    01/06/2013 12:32:00
-//       01 June 2013 12:32 PM                 01/06/2013 12:32:00
-//       01 June 2013 12:32 PM                 01/06/2013 12:32:00
-//       1 June 2013 12:32                     01/06/2013 12:32:00
-//       1 June 2013 12:32                     01/06/2013 12:32:00
-//       1 June 2013 12:32 PM                  01/06/2013 12:32:00
-//       1 June 2013 12:32 PM                  01/06/2013 12:32:00
-//       01 June 2013 12:32:30                 01/06/2013 12:32:30
-//       01 June 2013 12:32:30                 01/06/2013 12:32:30
-//       01 June 2013 12:32:30 PM              01/06/2013 12:32:30
-//       01 June 2013 12:32:30 PM              01/06/2013 12:32:30
-//       1 June 2013 12:32:30                  01/06/2013 12:32:30
-//       1 June 2013 12:32:30                  01/06/2013 12:32:30
-//       1 June 2013 12:32:30 PM               01/06/2013 12:32:30
-//       1 June 2013 12:32:30 PM               01/06/2013 12:32:30
-//       01/06/2013 12:32                      01/06/2013 12:32:00
-//       01/06/2013 12:32                      01/06/2013 12:32:00
-//       01/06/2013 12:32 PM                   01/06/2013 12:32:00
-//       01/06/2013 12:32 PM                   01/06/2013 12:32:00
-//       01/06/13 12:32                        01/06/2013 12:32:00
-//       01/06/13 12:32                        01/06/2013 12:32:00
-//       01/06/13 12:32 PM                     01/06/2013 12:32:00
-//       01/06/13 12:32 PM                     01/06/2013 12:32:00
-//       1/6/13 12:32                          01/06/2013 12:32:00
-//       1/6/13 12:32                          01/06/2013 12:32:00
-//       1/6/13 12:32 PM                       01/06/2013 12:32:00
-//       1/6/13 12:32 PM                       01/06/2013 12:32:00
-//       1.6.13 12:32                          01/06/2013 12:32:00
-//       1.6.13 12:32                          01/06/2013 12:32:00
-//       1.6.13 12:32 PM                       01/06/2013 12:32:00
-//       1.6.13 12:32 PM                       01/06/2013 12:32:00
-//       2013-06-01 12:32                      01/06/2013 12:32:00
-//       2013-06-01 12:32                      01/06/2013 12:32:00
-//       2013-06-01 12:32 PM                   01/06/2013 12:32:00
-//       2013-06-01 12:32 PM                   01/06/2013 12:32:00
-//       01/06/2013 12:32:30                   01/06/2013 12:32:30
-//       01/06/2013 12:32:30                   01/06/2013 12:32:30
-//       01/06/2013 12:32:30 PM                01/06/2013 12:32:30
-//       01/06/2013 12:32:30 PM                01/06/2013 12:32:30
-//       01/06/13 12:32:30                     01/06/2013 12:32:30
-//       01/06/13 12:32:30                     01/06/2013 12:32:30
-//       01/06/13 12:32:30 PM                  01/06/2013 12:32:30
-//       01/06/13 12:32:30 PM                  01/06/2013 12:32:30
-//       1/6/13 12:32:30                       01/06/2013 12:32:30
-//       1/6/13 12:32:30                       01/06/2013 12:32:30
-//       1/6/13 12:32:30 PM                    01/06/2013 12:32:30
-//       1/6/13 12:32:30 PM                    01/06/2013 12:32:30
-//       1.6.13 12:32:30                       01/06/2013 12:32:30
-//       1.6.13 12:32:30                       01/06/2013 12:32:30
-//       1.6.13 12:32:30 PM                    01/06/2013 12:32:30
-//       1.6.13 12:32:30 PM                    01/06/2013 12:32:30
-//       2013-06-01 12:32:30                   01/06/2013 12:32:30
-//       2013-06-01 12:32:30                   01/06/2013 12:32:30
-//       2013-06-01 12:32:30 PM                01/06/2013 12:32:30
-//       2013-06-01 12:32:30 PM                01/06/2013 12:32:30
-//       01 June                               01/06/2013 00:00:00
-//       01 June                               01/06/2013 00:00:00
-//       2013-06-01T12:32:30.0000000           01/06/2013 12:32:30
-//       2013-06-01T12:32:30.0000000           01/06/2013 12:32:30
-//       Sat, 01 Jun 2013 12:32:30 GMT         01/06/2013 05:32:30
-//       Sat, 01 Jun 2013 12:32:30 GMT         01/06/2013 05:32:30
-//       2013-06-01T12:32:30                   01/06/2013 12:32:30
-//       12:32                                 22/04/2013 12:32:00
-//       12:32                                 22/04/2013 12:32:00
-//       12:32 PM                              22/04/2013 12:32:00
-//       12:32 PM                              22/04/2013 12:32:00
-//       12:32:30                              22/04/2013 12:32:30
-//       12:32:30                              22/04/2013 12:32:30
-//       12:32:30 PM                           22/04/2013 12:32:30
-//       12:32:30 PM                           22/04/2013 12:32:30
-//       2013-06-01 12:32:30Z                  01/06/2013 05:32:30
-//       01 June 2013 19:32:30                 01/06/2013 19:32:30
-//       01 June 2013 19:32:30                 01/06/2013 19:32:30
-//       01 June 2013 07:32:30 PM              01/06/2013 19:32:30
-//       01 June 2013 7:32:30 PM               01/06/2013 19:32:30
-//       1 June 2013 19:32:30                  01/06/2013 19:32:30
-//       1 June 2013 19:32:30                  01/06/2013 19:32:30
-//       1 June 2013 07:32:30 PM               01/06/2013 19:32:30
-//       1 June 2013 7:32:30 PM                01/06/2013 19:32:30
-//       June 2013                             01/06/2013 00:00:00
-//       June 2013                             01/06/2013 00:00:00
-// </Snippet1>
-      }
-   }
-}
