+++ samples/snippets/visualbasic/System.DateTime/Program.vb
@@ -1,0 +1,12 @@
+Imports System
+
+Module Program
+    Sub Main(args As String())
+        Instantiation.Snippets()
+        StringFormat.Snippets()
+        Parsing.Snippets()
+        Resolution.Snippets()
+        CalendarSamples.Snippets()
+        Persistence.Snippets()
+    End Sub
+End Module
