--- samples/snippets/visualbasic/VS_Snippets_CLR_System/system.datetime.parsing/vb/parsingex1.vb
@@ -1,129 +1,0 @@
-' Visual Basic .NET Document
-Option Strict On
-
-Imports System.Collections.Generic
-Imports System.Globalization
-Imports System.Threading
-
-Module Example
-   Public Sub Main()
-' <Snippet1>
-      Thread.CurrentThread.CurrentCulture = CultureInfo.CreateSpecificCulture("en-GB")
-
-      Dim date1 As DateTime = New DateTime(2013, 6, 1, 12, 32, 30)
-      Dim badFormats As New List(Of String)
-      
-      Console.WriteLine($"{"Date String",-37} {"Date",-19}")
-      Console.WriteLine()
-      For Each dateString As String In date1.GetDateTimeFormats()
-         Dim parsedDate As DateTime
-         If DateTime.TryParse(dateString, parsedDate) Then
-            Console.WriteLine($"{dateString,-37} {DateTime.Parse(dateString),-19:g}")
-         Else
-            badFormats.Add(dateString)
-         End If   
-      Next 
-      
-      ' Display strings that could not be parsed.
-      If badFormats.Count > 0 Then
-         Console.WriteLine()
-         Console.WriteLine("Strings that could not be parsed: ")
-         For Each badFormat In badFormats
-            Console.WriteLine($"   {badFormat}")         
-         Next
-      End If
-' The example displays the following output:
-'       Date String                           Date               
-'       
-'       01/06/2013                            01/06/2013 00:00:00
-'       01/06/13                              01/06/2013 00:00:00
-'       1/6/13                                01/06/2013 00:00:00
-'       1.6.13                                01/06/2013 00:00:00
-'       2013-06-01                            01/06/2013 00:00:00
-'       01 June 2013                          01/06/2013 00:00:00
-'       1 June 2013                           01/06/2013 00:00:00
-'       01 June 2013 12:32                    01/06/2013 12:32:00
-'       01 June 2013 12:32                    01/06/2013 12:32:00
-'       01 June 2013 12:32 PM                 01/06/2013 12:32:00
-'       01 June 2013 12:32 PM                 01/06/2013 12:32:00
-'       1 June 2013 12:32                     01/06/2013 12:32:00
-'       1 June 2013 12:32                     01/06/2013 12:32:00
-'       1 June 2013 12:32 PM                  01/06/2013 12:32:00
-'       1 June 2013 12:32 PM                  01/06/2013 12:32:00
-'       01 June 2013 12:32:30                 01/06/2013 12:32:30
-'       01 June 2013 12:32:30                 01/06/2013 12:32:30
-'       01 June 2013 12:32:30 PM              01/06/2013 12:32:30
-'       01 June 2013 12:32:30 PM              01/06/2013 12:32:30
-'       1 June 2013 12:32:30                  01/06/2013 12:32:30
-'       1 June 2013 12:32:30                  01/06/2013 12:32:30
-'       1 June 2013 12:32:30 PM               01/06/2013 12:32:30
-'       1 June 2013 12:32:30 PM               01/06/2013 12:32:30
-'       01/06/2013 12:32                      01/06/2013 12:32:00
-'       01/06/2013 12:32                      01/06/2013 12:32:00
-'       01/06/2013 12:32 PM                   01/06/2013 12:32:00
-'       01/06/2013 12:32 PM                   01/06/2013 12:32:00
-'       01/06/13 12:32                        01/06/2013 12:32:00
-'       01/06/13 12:32                        01/06/2013 12:32:00
-'       01/06/13 12:32 PM                     01/06/2013 12:32:00
-'       01/06/13 12:32 PM                     01/06/2013 12:32:00
-'       1/6/13 12:32                          01/06/2013 12:32:00
-'       1/6/13 12:32                          01/06/2013 12:32:00
-'       1/6/13 12:32 PM                       01/06/2013 12:32:00
-'       1/6/13 12:32 PM                       01/06/2013 12:32:00
-'       1.6.13 12:32                          01/06/2013 12:32:00
-'       1.6.13 12:32                          01/06/2013 12:32:00
-'       1.6.13 12:32 PM                       01/06/2013 12:32:00
-'       1.6.13 12:32 PM                       01/06/2013 12:32:00
-'       2013-06-01 12:32                      01/06/2013 12:32:00
-'       2013-06-01 12:32                      01/06/2013 12:32:00
-'       2013-06-01 12:32 PM                   01/06/2013 12:32:00
-'       2013-06-01 12:32 PM                   01/06/2013 12:32:00
-'       01/06/2013 12:32:30                   01/06/2013 12:32:30
-'       01/06/2013 12:32:30                   01/06/2013 12:32:30
-'       01/06/2013 12:32:30 PM                01/06/2013 12:32:30
-'       01/06/2013 12:32:30 PM                01/06/2013 12:32:30
-'       01/06/13 12:32:30                     01/06/2013 12:32:30
-'       01/06/13 12:32:30                     01/06/2013 12:32:30
-'       01/06/13 12:32:30 PM                  01/06/2013 12:32:30
-'       01/06/13 12:32:30 PM                  01/06/2013 12:32:30
-'       1/6/13 12:32:30                       01/06/2013 12:32:30
-'       1/6/13 12:32:30                       01/06/2013 12:32:30
-'       1/6/13 12:32:30 PM                    01/06/2013 12:32:30
-'       1/6/13 12:32:30 PM                    01/06/2013 12:32:30
-'       1.6.13 12:32:30                       01/06/2013 12:32:30
-'       1.6.13 12:32:30                       01/06/2013 12:32:30
-'       1.6.13 12:32:30 PM                    01/06/2013 12:32:30
-'       1.6.13 12:32:30 PM                    01/06/2013 12:32:30
-'       2013-06-01 12:32:30                   01/06/2013 12:32:30
-'       2013-06-01 12:32:30                   01/06/2013 12:32:30
-'       2013-06-01 12:32:30 PM                01/06/2013 12:32:30
-'       2013-06-01 12:32:30 PM                01/06/2013 12:32:30
-'       01 June                               01/06/2013 00:00:00
-'       01 June                               01/06/2013 00:00:00
-'       2013-06-01T12:32:30.0000000           01/06/2013 12:32:30
-'       2013-06-01T12:32:30.0000000           01/06/2013 12:32:30
-'       Sat, 01 Jun 2013 12:32:30 GMT         01/06/2013 05:32:30
-'       Sat, 01 Jun 2013 12:32:30 GMT         01/06/2013 05:32:30
-'       2013-06-01T12:32:30                   01/06/2013 12:32:30
-'       12:32                                 22/04/2013 12:32:00
-'       12:32                                 22/04/2013 12:32:00
-'       12:32 PM                              22/04/2013 12:32:00
-'       12:32 PM                              22/04/2013 12:32:00
-'       12:32:30                              22/04/2013 12:32:30
-'       12:32:30                              22/04/2013 12:32:30
-'       12:32:30 PM                           22/04/2013 12:32:30
-'       12:32:30 PM                           22/04/2013 12:32:30
-'       2013-06-01 12:32:30Z                  01/06/2013 05:32:30
-'       01 June 2013 19:32:30                 01/06/2013 19:32:30
-'       01 June 2013 19:32:30                 01/06/2013 19:32:30
-'       01 June 2013 07:32:30 PM              01/06/2013 19:32:30
-'       01 June 2013 7:32:30 PM               01/06/2013 19:32:30
-'       1 June 2013 19:32:30                  01/06/2013 19:32:30
-'       1 June 2013 19:32:30                  01/06/2013 19:32:30
-'       1 June 2013 07:32:30 PM               01/06/2013 19:32:30
-'       1 June 2013 7:32:30 PM                01/06/2013 19:32:30
-'       June 2013                             01/06/2013 00:00:00
-'       June 2013                             01/06/2013 00:00:00
-' </Snippet1>
-   End Sub
-End Module
